Fixed edit and delete recipes
Fixed edit recipes - now it is working
Fixed delete recipes - now it is working
Edited some pure js to jquery for simplier syntax

+++ src/app/Controllers/AjaxHandler.php
@@ -7,11 +7,19 @@ use App\Models\RecipeIngredientsModel;
 use CodeIgniter\Exceptions\PageNotFoundException;
 
 class AjaxHandler extends BaseController{
-	public function index($recipe_id=null){
+	public function index(){
+		if (! $this->validate([
+            'id' => 'required'
+		])) {
+            // The validation fails, so returns the form.
+            throw new PageNotFoundException('Zle zadaný request');;
+        }
+		$recipe_id = ($this->validator->getValidated())['id'];
         $model = model(RecipeModel::class);
         $model_steps = model(RecipeStepsModel::class);
         $model_ingredients = model(RecipeIngredientsModel::class);
         $recipe = $model->getRecipes($recipe_id,null);
+		$data['recipe_types']=['polievkové lyžice','čajové lyžice','kusy','balenia','kilogramy','gramy','mililitre','decilitre','litre','trochu','veľa'];
 		$data['id']=$recipe['id'];
         $data['recipe_name']=$recipe['recipe_name'];
         if (empty($data['recipe_name'])) {
@@ -19,24 +27,27 @@ class AjaxHandler extends BaseController{
         }
         $data['recipe_img_path']=$recipe["recipe_img_path"];
         $data['recipe_steps']=$model_steps->getRecipeSteps($recipe_id);
-        $data['recipe_ingredients']=$model_ingredients->getRecipeIngredients($recipe_id);
-		
-        return view('ajaxHandlers/editForm');
+        $data['recipe_ingredients']=$model_ingredients->getRecipeIngredients($recipe_id);		
+        return view('ajaxHandlers/editForm',$data);
     }
 	public function editRecipe(){
 		helper('form');
         //|regex_match[[^\\s]+(.*?)\\.(jpg|png|pneg)$]
+		
         if (! $this->validate([
+			'id' => 'required',
             'recipe_name' => 'required|max_length[32]|min_length[2]',
             'recipe_img_path'  => 'required|max_length[32]|min_length[2]',
             'recipe_steps' => 'required',
+			'recipe_steps_id' => 'required',
+			'recipe_ingredients_id' => 'required',
             'recipe_ingredient_names' => 'required',
             'recipe_ingredient_counts'  => 'required',
             'recipe_ingredient_types' => 'required'
         ])) {
             // The validation fails, so returns the form.
-			echo 'Something is wrong with input datas'
-            return ;
+			
+            return 'Something is wrong with inputed datas';
         }
 
         $model = model(RecipeModel::class);
@@ -48,13 +59,11 @@ class AjaxHandler extends BaseController{
         //if recipe ingredient info count not fit return form
         if (!(count($post['recipe_ingredient_names']) == count($post['recipe_ingredient_counts']) and
             count($post['recipe_ingredient_names']) == count($post['recipe_ingredient_types']))){
-            echo 'Count of ingredient infos are not same';
-            return ;
+            return 'Count of ingredient infos are not same';
         }
 		$recipe=$model->getRecipes($post['id']);
 		if ($recipe['user_id'] != session()->get('id')){
-			echo 'Dont change other recipes';
-            return ;
+            return 'Dont change other recipes';
 		}
 			
         $model->update($post['id'],[
@@ -64,8 +73,8 @@ class AjaxHandler extends BaseController{
         $i=1;
 		$model_steps->getRecipeSteps($post['id']);
         foreach ($post['recipe_steps'] as $recipe_step){
-			if ($recipe_step['step_id'] != null)
-				$model_steps->update($recipe_step['step_id'],[
+			if ($post['recipe_steps_id'][$i-1] != null)
+				$model_steps->update($post['recipe_steps_id'][$i-1],[
 					'step_description' => $recipe_step
 				]);
 			else $model_steps->insert([
@@ -75,8 +84,8 @@ class AjaxHandler extends BaseController{
             ]);
         }
         for ($i=0;$i<count($post['recipe_ingredient_names']);$i++) {
-			if ($recipe_step['ingredient_id'] != null)
-				$model_ingredients->update($post['ingredient_id'][$i],[
+			if ($post['recipe_ingredients_id'][$i] != null)
+				$model_ingredients->update($post['recipe_ingredients_id'][$i],[
 					'ingredient_name' => $post['recipe_ingredient_names'][$i],
 					'ingredient_count' => $post['recipe_ingredient_counts'][$i],
 					'ingredient_count_type' => $post['recipe_ingredient_types'][$i],
@@ -89,18 +98,24 @@ class AjaxHandler extends BaseController{
 					'ingredient_count_type' => $post['recipe_ingredient_types'][$i],
 				]);
         }
-		echo 'Succesful edited recipe '.$post['recipe_name'].'';
+		return json_encode(["scs" => true, "msg" => '<h2 class="green">Succesful edited recipe '.$post['recipe_name'].'</h2>']);
 	}
 	public function deleteRecipe(){
+		if (! $this->validate([
+            'id' => 'required'
+		])) {
+            // The validation fails, so returns the form.
+            throw new PageNotFoundException('Zle zadaný request');;
+        }
+		$post = $this->validator->getValidated();
 		$model = model(RecipeModel::class);
-		$your_id=isset(session()->get('id')) ? session()->get('id') : null;
+		$your_id=session()->get('id') !==null ? session()->get('id') : null;
 		$recipe=$model->getRecipes($post['id']);
 		if ($your_id && $your_id==$recipe['user_id']){
 			$model->where('id', $recipe['id'])->delete();
-			echo json_encode(["scs" => true, "msg" => '<h2 class="blue">Sucessful deleted recipe</h2>']);
+			return json_encode(["scs" => true, "msg" => '<h2 class="green">Sucessful deleted recipe</h2>']);
 		}
 		else
-			echo json_encode(["scs" => false, "msg" => '<h2 class="red">Unsucessful attempt to delete recipe</h2>']);		
+			return json_encode(["scs" => false, "msg" => '<h2 class="red">Unsucessful attempt to delete recipe</h2>']);		
 	}
-	
 }
